// generated/components/vikAmericasLevels.ts
export type Tile = { f: string; x0: number; y0: number; w: number; h: number; W: number; H: number };
export type Level = {
  name: string;
  kBake: number;
  s: number;
  band: [number, number] | null;
  x0: number;
  y0: number;
  w: number;
  h: number;
  W: number;
  H: number;
  tiles: Tile[];
};
export const LEVELS: Level[] = [
 {
  "name": "far",
  "kBake": 1,
  "s": 1.5,
  "band": null,
  "x0": -355,
  "y0": -449,
  "w": 1810,
  "h": 2798,
  "W": 2715,
  "H": 4197,
  "tiles": [
   {
    "f": "far-0-0.png",
    "x0": -355,
    "y0": -449,
    "w": 1366.6666666666667,
    "h": 1366.6666666666667,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "far-1-0.png",
    "x0": 1010.3333333333333,
    "y0": -449,
    "w": 444.6666666666667,
    "h": 1366.6666666666667,
    "W": 667,
    "H": 2050
   },
   {
    "f": "far-0-1.png",
    "x0": -355,
    "y0": 916.3333333333333,
    "w": 1366.6666666666667,
    "h": 1366.6666666666667,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "far-1-1.png",
    "x0": 1010.3333333333333,
    "y0": 916.3333333333333,
    "w": 444.6666666666667,
    "h": 1366.6666666666667,
    "W": 667,
    "H": 2050
   },
   {
    "f": "far-0-2.png",
    "x0": -355,
    "y0": 2281.6666666666665,
    "w": 1366.6666666666667,
    "h": 67.33333333333333,
    "W": 2050,
    "H": 101
   },
   {
    "f": "far-1-2.png",
    "x0": 1010.3333333333333,
    "y0": 2281.6666666666665,
    "w": 444.6666666666667,
    "h": 67.33333333333333,
    "W": 667,
    "H": 101
   }
  ]
 },
 {
  "name": "near",
  "kBake": 2,
  "s": 2.75,
  "band": [
   1.3,
   1.45
  ],
  "x0": -659,
  "y0": -569,
  "w": 2398.181818181818,
  "h": 2904,
  "W": 6595,
  "H": 7986,
  "tiles": [
   {
    "f": "near-0-0.png",
    "x0": -659,
    "y0": -569,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-1-0.png",
    "x0": 85.72727272727275,
    "y0": -569,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-2-0.png",
    "x0": 830.4545454545455,
    "y0": -569,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-3-0.png",
    "x0": 1575.181818181818,
    "y0": -569,
    "w": 164,
    "h": 745.4545454545455,
    "W": 451,
    "H": 2050
   },
   {
    "f": "near-0-1.png",
    "x0": -659,
    "y0": 175.72727272727275,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-1-1.png",
    "x0": 85.72727272727275,
    "y0": 175.72727272727275,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-2-1.png",
    "x0": 830.4545454545455,
    "y0": 175.72727272727275,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-3-1.png",
    "x0": 1575.181818181818,
    "y0": 175.72727272727275,
    "w": 164,
    "h": 745.4545454545455,
    "W": 451,
    "H": 2050
   },
   {
    "f": "near-0-2.png",
    "x0": -659,
    "y0": 920.4545454545455,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-1-2.png",
    "x0": 85.72727272727275,
    "y0": 920.4545454545455,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-2-2.png",
    "x0": 830.4545454545455,
    "y0": 920.4545454545455,
    "w": 745.4545454545455,
    "h": 745.4545454545455,
    "W": 2050,
    "H": 2050
   },
   {
    "f": "near-3-2.png",
    "x0": 1575.181818181818,
    "y0": 920.4545454545455,
    "w": 164,
    "h": 745.4545454545455,
    "W": 451,
    "H": 2050
   },
   {
    "f": "near-0-3.png",
    "x0": -659,
    "y0": 1665.181818181818,
    "w": 745.4545454545455,
    "h": 669.8181818181819,
    "W": 2050,
    "H": 1842
   },
   {
    "f": "near-1-3.png",
    "x0": 85.72727272727275,
    "y0": 1665.181818181818,
    "w": 745.4545454545455,
    "h": 669.8181818181819,
    "W": 2050,
    "H": 1842
   },
   {
    "f": "near-2-3.png",
    "x0": 830.4545454545455,
    "y0": 1665.181818181818,
    "w": 745.4545454545455,
    "h": 669.8181818181819,
    "W": 2050,
    "H": 1842
   },
   {
    "f": "near-3-3.png",
    "x0": 1575.181818181818,
    "y0": 1665.181818181818,
    "w": 164,
    "h": 669.8181818181819,
    "W": 451,
    "H": 1842
   }
  ]
 }
];